requirements.txt
absl-py                 2.1.0
asttokens               3.0.0
astunparse              1.6.3
certifi                 2024.12.14
charset-normalizer      3.4.0
colorama                0.4.6
comm                    0.2.2
contourpy               1.3.1
cycler                  0.12.1
debugpy                 1.8.9
decorator               5.1.1
executing               2.1.0
filelock                3.16.1
flatbuffers             24.3.25
fonttools               4.55.2
fsspec                  2024.10.0
gast                    0.6.0
google-pasta            0.2.0
grpcio                  1.68.1
h5py                    3.12.1
idna                    3.10
intel-cmplr-lib-ur      2025.0.3
intel-openmp            2025.0.3
ipykernel               6.29.5
ipython                 8.30.0
jedi                    0.19.2
Jinja2                  3.1.4
jupyter_client          8.6.3
jupyter_core            5.7.2
keras                   3.7.0
kiwisolver              1.4.7
libclang                18.1.1
Markdown                3.7
markdown-it-py          3.0.0
MarkupSafe              3.0.2
matplotlib              3.9.3
matplotlib-inline       0.1.7
mdurl                   0.1.2
mkl                     2025.0.1
mkl-service             2.4.2
ml-dtypes               0.4.1
mpmath                  1.3.0
namex                   0.0.8
nest-asyncio            1.6.0
networkx                3.4.2
numpy                   2.0.2
opencv-python           4.10.0.84
opt_einsum              3.4.0
optree                  0.13.1
packaging               24.2
pandas                  2.2.3
parso                   0.8.4
pillow                  11.0.0
pip                     24.3.1
platformdirs            4.3.6
prompt_toolkit          3.0.48
protobuf                5.29.2
psutil                  6.1.0
pure_eval               0.2.3
pycocotools             2.0.8
Pygments                2.18.0
pyparsing               3.2.0
python-dateutil         2.9.0.post0
pytz                    2024.2
pywin32                 308
pyzmq                   26.2.0
requests                2.32.3
rich                    13.9.4
setuptools              75.6.0
six                     1.17.0
stack-data              0.6.3
sympy                   1.13.1
tbb                     2022.0.0
tcmlib                  1.2.0
tensorboard             2.18.0
tensorboard-data-server 0.7.2
tensorflow_intel        2.18.0
termcolor               2.5.0
torch                   2.5.1+cu121
torchvision             0.20.1+cu121
tornado                 6.4.2
traitlets               5.14.3
typing_extensions       4.12.2
tzdata                  2024.2
umf                     0.9.1
urllib3                 2.3.0
wcwidth                 0.2.13
Werkzeug                3.1.3
wheel                   0.45.1
wrapt                   1.17.0